+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 * ensure close() is properly called from the response (bubenkoff, GrahamDumpleton)
 * optionally time exceptions (bubenkoff)
+* code readability improvements (bubenkoff)
 
 0.2.0
 -----